requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
Automat==22.10.0
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.8
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.14
djangorestframework==3.15.2
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
ipdb==0.13.13
ipython==8.12.3
itemadapter==0.9.0
itemloaders==1.3.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
orjson==3.10.6
packaging==24.1
parsel==1.9.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt_toolkit==3.0.47
Protego==0.3.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
Scrapy==2.11.2
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.37.2
tldextract==5.1.2
tomli==2.0.1
traitlets==5.14.3
Twisted==24.3.0
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
w3lib==2.2.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
zope.interface==6.4.post2
